gtkui: don't fill bg when seeking
When clicking the gtk2ui seekbar it displays the new position in the
stream. However, it renders a filled background behind this text. This
background can be confused with the fill that corresponds to actual
progress in the stream. That is, seeking near the middle of the stream
makes it look as if a little more than the stream has passed. Or, if
seeking near the start of the stream, two discontinuous fills are seen.